Improve: Cleanup
 - Remove Yarn lock file (switched to NPM)
 - Fix type issue
 - Small refactor
